core/requirements.txt
asyncpg

minio

fastapi
pyjwt
pydantic[email]
python-multipart
uvicorn[standard]